complianceiq/requirements.txt
streamlit>=1.28.0
openai>=1.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
pytest>=7.0.0
